% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Return the windows and the SNPs with the higest sginificance scores}
\usage{
get_peaks(data_raw, data_sum, stat_col, score_col, score_th, window_s,
  greater = FALSE)
}
\arguments{
\item{data_raw:}{data.table with every SNP position}

\item{data_sum:}{data.table with the following columns: CHR, SNP, CM, POS, col_name}

\item{stat_col:}{Name of the stat column}

\item{score_col:}{Name of the score column}

\item{score_th:}{double with the score thereshold}

\item{window_s:}{integer with the size of the window used in the rolling
statistics}

\item{greater:}{BOOLEAN deafault TRUE}
}
\value{
data.table
}
\description{
\code{get_peaks()} Return the windows and the SNPs with the higest scores
}
