% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudson_fst.R
\name{hudson_fst}
\alias{hudson_fst}
\title{Estimate the FST using the Hudson formula}
\usage{
hudson_fst(pop_list)
}
\arguments{
\item{pop_list:}{list of where each element is a data.table with the
following columns: CHR, SNP, CM, POS, NCHROBS, POP, VAR, AF}
}
\value{
data.table with Fst estimated for every SNP
}
\description{
\code{wc_fst()} estimates the Fst for every variant
}
