% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bim_frq.R
\name{load_bim_frq}
\alias{load_bim_frq}
\title{Create data.table with CHR, SNP, POS, AF, POP columns}
\usage{
load_bim_frq(files_path, pop_name = tail(unlist(strsplit(files_path, "/")), n
  = 1), one_allele_perline = TRUE)
}
\arguments{
\item{files_path}{string with full path of .bim and .frq files}

\item{pop_name}{string with the name of the population. DEFAULT is the
prefix of .bim and .frq files.}

\item{one_allele_perline}{boolean. Transform data from one SNP per line to
one allele per line, Default TRUE.}
}
\value{
data.table with the information of .bim and .frq files
}
\description{
\code{load_bim_frq()} reads .bim and .frq plink files
and merge then into one data.table
}
