% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Draw a manhattan plot from data}
\usage{
manhattan_plot(DT_in, col_name, fig_name = NULL, facet_key = NULL)
}
\arguments{
\item{facet_key}{}

\item{data.table:}{has to have the columns CHR, POS, p.value}

\item{fig_name:}{figure name with extension (.png, .eps, .pdf): Defaut NULL}
}
\value{
ggplot
}
\description{
\code{manhattan_plot()} Create a manhatan plot
}
