% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_fst.R
\name{pbs_fst}
\alias{pbs_fst}
\title{Estimate the population branch size of for every SNP}
\usage{
pbs_fst(focal_pop, close_pop, out_pop)
}
\arguments{
\item{focal_pop:}{data.table with the following columns: CHR, SNP, CM, POS,
NCHROBS, POP, VAR, AF.}

\item{close_pop:}{data.table}

\item{out_pop:}{data.table}
}
\value{
data.table with the estimated branch size for every SNP
}
\description{
\code{pbs_fst()} estimates the branch size for every SNP
}
