% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_index_by_snp_number.R
\name{rolling_index_by_snp_number}
\alias{rolling_index_by_snp_number}
\title{Return the index based in the number of SNPs for rolling statiscs}
\usage{
rolling_index_by_snp_number(data_in, window_s = 3, step_s = 1)
}
\arguments{
\item{focal_pop:}{data.table with the following columns: CHR, SNP, CM, POS,
NCHROBS, POP, VAR, AF.}

\item{close_pop:}{data.table}

\item{out_pop:}{data.table}
}
\value{
data.table with the estimated branch size for every SNP
}
\description{
\code{rolling_index_by_snp_number()} estimates the branch size for every SNP
}
