% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_median_n.R
\name{rolling_median_n}
\alias{rolling_median_n}
\title{Evaluete the sliding window meadian}
\usage{
rolling_median_n(data_in, col_name, window_s = 3, step_s = 2)
}
\arguments{
\item{data_in:}{data.table with the following columns: CHR, POS, col_name}

\item{col_name:}{name of the columns to be take the mean}

\item{widows_s:}{number of SNPs in the rolling window}

\item{step_s:}{number of SNPs in the rolling window}
}
\value{
data.table with the rolling mean
}
\description{
\code{rolling_median_n()} evaluetes the rolling mean
}
